Add babel require hook to the serevr
